moved `View` struct to own module within `window`
