+++ Cargo.toml
@@ -5,7 +5,7 @@ version = "0.3.0"
 
 [[bin]]
 name = "gol-win"
-path = "./src/window.rs"
+path = "./src/window/mod.rs"
 
 [[bin]]
 name = "gol-cli"
